refactor: more basic card logic to deck
